sandbox/python-google-cloud/pyproject.toml
[tool.poetry]
name = "python-google-cloud"
version = "0.1.0"
description = ""
authors = ["Tsubasa Irisawa <lambdasawa@gmail.com>"]
readme = "README.md"
packages = [{include = "python_google_cloud"}]

[tool.poetry.dependencies]
python = "^3.11"
google-cloud-storage = "^2.9.0"
google-cloud-pubsub = "^2.17.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
